@@ -5,20 +5,20 @@ version: 3.5.2
 homepage: https://github.com/mo-ah-dawood/fcm_config/tree/master/fcm_config
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
+  sdk: ">=3.5.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^2.16.0
-  firebase_messaging: ^14.6.8
-  flutter_local_notifications: ^15.1.1
-  http: ^1.1.0
-  path_provider: ^2.1.1
+  firebase_core: ^3.3.0
+  firebase_messaging: ^15.0.4
+  flutter_local_notifications: ^17.2.2
+  http: ^1.2.2
+  path_provider: ^2.1.4
 
 dev_dependencies:
-  flutter_lints: ^2.0.3
+  flutter_lints: ^4.0.0
 
 false_secrets:
   - /example/android/app/google-services.json
