@@ -3,10 +3,10 @@ description: Browse our products with ease and choose your favorite meat cut and
 
 publish_to: 'none'
 
-version: 20.0.5+1
+version: 20.1.0+1
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,76 +15,87 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  cupertino_icons: ^1.0.6
-  auto_route: ^7.8.4
+  cupertino_icons: ^1.0.8
+  auto_route: ^9.2.2
   intl: ^0.19.0
-  flutter_riverpod: ^2.4.9
-  riverpod_annotation: ^2.3.3
-  flutter_native_splash: ^2.3.10
-  flutter_svg: ^2.0.9
-  google_maps_flutter: ^2.5.3
-  modal_bottom_sheet: ^3.0.0-pre
-  geocoding: ^2.1.1
-  uuid: ^4.2.1
-  dio: ^5.4.0
-  geolocator: ^10.1.0
-  mockito: ^5.4.3
-  cached_network_image: ^3.3.0
-  carousel_slider: ^4.2.1
+  flutter_riverpod: ^2.6.1
+  riverpod_annotation: ^2.6.1
+  flutter_native_splash: ^2.4.1
+  flutter_svg: ^2.0.10
+  google_maps_flutter: ^2.9.0
+  modal_bottom_sheet: ^3.0.0
+  geocoding: ^3.0.0
+  uuid: ^4.5.1
+  dio: ^5.7.0
+  geolocator: ^13.0.1
+  mockito: ^5.4.4
+  cached_network_image: ^3.4.1
+  carousel_slider: ^5.0.0
   dots_indicator: ^3.0.0
   pull_to_refresh: ^2.0.0
   queen_validators: ^1.0.1  
   flutter_staggered_grid_view: ^0.7.0
-  pinput: ^3.0.1
+  pinput: ^5.0.0
   timer_count_down: ^2.2.2
-  easy_stepper: ^0.8.2
+  easy_stepper: ^0.8.5
   flutter_keyboard_visibility: ^6.0.0
   dotted_border: ^2.1.0
-  shared_preferences: ^2.2.2
-  image_picker: ^1.0.7
-  freezed_annotation: ^2.4.1
-  json_annotation: ^4.8.1
-  url_launcher: ^6.2.4
-  package_info_plus: ^5.0.1
-  webview_flutter: ^4.4.4
-  pay: ^1.1.2
+  shared_preferences: ^2.3.2
+  image_picker: ^1.1.2
+  freezed_annotation: ^2.4.4
+  json_annotation: ^4.9.0
+  url_launcher: ^6.3.0
+  package_info_plus: ^8.0.2
+  webview_flutter: ^4.9.0
+  pay: ^2.0.0
   fcm_config: 
     path: local_packages/fcm_config
-  store_redirect: ^2.0.1
+  store_redirect:
+    git:
+      url: https://github.com/Danesz/store_redirect
+      ref: 8edbefa83bf3f276538c4b4464c10849585db743
   flutter_html: ^3.0.0-beta.2
-  flutter_barcode_scanner: ^2.0.0
-  firebase_crashlytics: ^3.4.9
-  firebase_core: ^2.24.2
-  onboarding_overlay: ^3.2.1
+  flutter_barcode_scanner:
+    git:
+      url: https://github.com/AmolGangadhare/flutter_barcode_scanner
+      ref: 379ded73012f1221a372edba8867ac3341b27f9e
+  firebase_crashlytics: ^4.1.3
+  firebase_core: ^3.6.0
+  onboarding_overlay: ^3.2.2
   flutter_rating_bar: ^4.0.1
-  widget_to_marker: ^1.0.5
+  widget_to_marker: ^1.0.6
   flutter_polyline_points_plus: ^3.0.0
   step_progress_indicator: ^1.0.2
-  synchronized: ^3.1.0+1
-  flutter_vibrate: ^1.3.0
+  synchronized: ^3.3.0
   flutter_cupertino_date_picker_fork: ^1.0.7
-  firebase_database: ^10.4.0
-  firebase_ui_database: ^1.4.1
-  firebase_auth: ^4.16.0
+  firebase_database: ^11.1.4
+  firebase_ui_database: ^1.5.0
+  firebase_auth: ^5.3.1
   status_bar_control: ^3.2.1
   flutter_chat_bubble: ^2.0.2
-  firebase_analytics: ^10.8.0
-  map_launcher: ^3.3.0
+  firebase_analytics: ^11.3.3
+  map_launcher: ^3.5.0
+  smart_auth: ^2.0.0
+  flutter_vibrate:
+    path: local_packages/flutter_vibrate-master
+
+dependency_overrides:
+  web: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.1
-  auto_route_generator: ^7.3.2
-  build_runner: ^2.4.8
-  riverpod_generator: ^2.3.9
-  flutter_gen_runner: ^5.4.0
-  riverpod_lint: ^2.3.7
-  custom_lint: ^0.5.11
-  freezed: ^2.4.6
-  json_serializable: ^6.7.1
-  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^5.0.0
+  auto_route_generator: ^9.0.0
+  build_runner: ^2.4.13
+  riverpod_generator: ^2.6.2
+  flutter_gen_runner: ^5.7.0
+  riverpod_lint: ^2.6.2
+  custom_lint: ^0.7.0
+  freezed: ^2.5.7
+  json_serializable: ^6.8.0
+  flutter_launcher_icons: ^0.14.1
 
 flutter_launcher_icons:
   android: false
