@@ -1,5 +1,5 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import '../../../application/geofence_id_controller.dart';
+import '../../../application/local_location_info.dart';
 
 import '../../../data/addresses_repository.dart';
 import '../../../domain/address/address.dart';
@@ -15,7 +15,7 @@ class AddressesListController extends _$AddressesListController {
 
   void addAddress(Address address) {
     state = AsyncData([...state.asData!.value, address]);
-    _refreshAppGeofenceId(address);
+    _refreshAppLocationInfo(address);
   }
 
   void updateAddress(Address address) {
@@ -23,23 +23,24 @@ class AddressesListController extends _$AddressesListController {
       for (final item in state.asData!.value)
         if (item.addressId == address.addressId) address else item
     ]);
-    _refreshAppGeofenceId(address);
+    _refreshAppLocationInfo(address);
   }
 
   void deleteAddress(Address address) {
     final defaultAddress = state.asData!.value
         .firstWhere((element) => element.preferredShippingAddress == 1);
-    _refreshAppGeofenceId(defaultAddress);
+    _refreshAppLocationInfo(defaultAddress);
     state = AsyncData([
       for (final item in state.asData!.value)
         if (item.addressId != address.addressId) item
     ]);
   }
 
-  void _refreshAppGeofenceId(Address address) {
+  void _refreshAppLocationInfo(Address address) {
+    ref.read(localLocationInfoProvider.notifier).setLocalLocationInfo(
+        address.latitude, address.longitude, address.warehouse?.warehouseId);
     ref
-        .read(geofenceIdAndCoordinatesProvider.notifier)
-        .setGeofenceIdAndCoordinates(
-            address.geofence!.geofenceId, address.latitude, address.longitude);
+        .read(localGeofenceIdProvider.notifier)
+        .setLocalGeofenceIdWithoutUpdateState(address.geofence!.geofenceId);
   }
 }
