@@ -1,5 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import '../../addresses/application/local_location_info.dart';
 import '../data/item_details_repository.dart';
 import '../domain/item_details/item_details.dart';
 
@@ -10,33 +12,37 @@ class ItemDetailsService {
 
   const ItemDetailsService(this._itemDetailsRepository);
 
-  Future<ItemDetails> getItemDetails({
-    required String itemId,
-    String? mubadaraId,
-    String? attributionToken,
-  }) async {
+  Future<ItemDetails> getItemDetails(
+      {required String itemId,
+      String? mubadaraId,
+      String? attributionToken,
+      required Ref ref}) async {
     final itemDetails = await _getItemDetails(
         itemId: itemId,
         mubadaraId: mubadaraId,
-        attributionToken: attributionToken);
+        attributionToken: attributionToken,
+        ref: ref);
     return itemDetails;
   }
 
   Future<ItemDetails> _getItemDetails(
       {required String itemId,
       String? mubadaraId,
-      String? attributionToken}) async {
+      String? attributionToken,
+      required Ref ref}) async {
+    final warehouseId = ref.watch(localLocationInfoProvider).warehouseId;
     final repository = _itemDetailsRepository;
     return repository.getItemDetails(
         itemId: itemId,
         mubadaraId: mubadaraId,
-        attributionToken: attributionToken);
+        attributionToken: attributionToken,
+        warehouseId: warehouseId);
   }
 
   Future<ItemDetails> onVariantsChange(
-      {required String itemId, String? mubadaraId}) async {
+      {required String itemId, String? mubadaraId, required Ref ref}) async {
     final itemDetails =
-        await _getItemDetails(itemId: itemId, mubadaraId: mubadaraId);
+        await _getItemDetails(itemId: itemId, mubadaraId: mubadaraId, ref: ref);
     return itemDetails;
   }
 }
