@@ -1,5 +1,6 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:widam/src/features/subscriptions/presentation/subscription_screen/subscription_body/set_your_schedule/subscription_info/subscription_title_controller.dart';
+import '../../../../../addresses/application/local_location_info.dart';
 import '../../../../data/subscriptions_repository.dart';
 import '../set_your_schedule/subscription_info/subscription_info_controller.dart';
 
@@ -18,8 +19,10 @@ class SendSubscriptionAsync extends _$SendSubscriptionAsync {
       final subscriptionInfo = ref.watch(subscriptionInfoControllerProvider);
       final newSubscriptionInfo = subscriptionInfo.copyWith(
           title: ref.watch(subscriptionTitleControllerProvider));
-      final appResponse =
-          await subscriptionsRepository.addSubscription(newSubscriptionInfo);
+      final warehouseId =
+          ref.watch(localLocationInfoProvider).warehouseId;
+      final appResponse = await subscriptionsRepository.addSubscription(
+          newSubscriptionInfo, warehouseId);
       state = AsyncValue.data(appResponse.message);
     } catch (e, stack) {
       state = AsyncValue.error(e, stack);
