@@ -12,7 +12,8 @@ class TimeSlotLabel extends StatelessWidget {
       required this.formattedDate,
       required this.formattedTime,
       this.foregroundColor = Colors.black,
-      this.onTap, this.customeTitle});
+      this.onTap,
+      this.customeTitle});
   final String formattedDate;
   final String formattedTime;
   final VoidCallback? onTap;
@@ -25,7 +26,9 @@ class TimeSlotLabel extends StatelessWidget {
         TruckTimeIcon(foregroundColor),
         const SizedBox(width: 8),
         EarliestDeliveryText(
-            isChangeable: onTap != null, foregroundColor: foregroundColor, customeTitle: customeTitle),
+            isChangeable: onTap != null,
+            foregroundColor: foregroundColor,
+            customeTitle: customeTitle),
         const SizedBox(width: 8),
         TimeSlotContainer(
           formattedDate: formattedDate,
@@ -41,17 +44,19 @@ class EarliestDeliveryText extends StatelessWidget {
   const EarliestDeliveryText({
     super.key,
     this.isChangeable = false,
-    this.foregroundColor = Colors.black, this.customeTitle,
+    this.foregroundColor = Colors.black,
+    this.customeTitle,
   });
   final bool isChangeable;
   final Color foregroundColor;
   final String? customeTitle;
   @override
   Widget build(BuildContext context) {
-    return Text(customeTitle ??
-        (isChangeable
-            ? S.of(context).changeTimeSlot
-            : S.of(context).earliestDelivery),
+    return Text(
+        customeTitle ??
+            (isChangeable
+                ? S.of(context).changeTimeSlot
+                : S.of(context).earliestDelivery),
         style: TextStyle(
             color: foregroundColor,
             fontSize: 11,
